refactor: product-context
