+++ src/Reducers/index.js
@@ -2,4 +2,5 @@ import { signupErrorReducer } from "./SignupErrorReducer";
 import { SignupUserReducer } from "./SignupUserReducer";
 import { cartReducerFunction } from "./cartReducer";
 import { wishlistReducerFunction } from "./wishlistReducer";
-export { signupErrorReducer, SignupUserReducer, cartReducerFunction, wishlistReducerFunction };
+import { productReducerFun } from "./ProductReducer";
+export { signupErrorReducer, SignupUserReducer, cartReducerFunction, wishlistReducerFunction, productReducerFun };
